ui/requirements.txt
streamlit==1.37.0
load_dotenv==0.1.0
nest_asyncio==1.6.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
serpapi
google-genai